Fix link type for test and init null buffer on pcap.next(..)

+++ jdk7/src/main/java/pcap/jdk7/internal/NativeMappings.java
@@ -317,7 +317,7 @@ class NativeMappings {
   @NativeSignature(
       signature = "u_int bpf_filter(const struct bpf_insn *, const u_char *, u_int, u_int)",
       since = @Version(major = 0, minor = 4))
-  static native int bpf_filter(
+  static native long bpf_filter(
       NativeMappings.bpf_insn insn, Pointer packet, int oriPktLen, int pktLen);
 
   static InetAddress inetAddress(sockaddr sockaddr) {
